+++ apps/hyperlocalise-web/src/lib/tools/file-tools.ts
@@ -5,33 +5,9 @@ import { and, eq } from "drizzle-orm";
 
 import { schema } from "@/lib/database";
 import { getFileStorageAdapter } from "@/lib/file-storage";
+import { bufferFromStream } from "@/lib/streams";
 
 import type { ToolContext } from "./types";
-
-function bufferFromStream(stream: ReadableStream<Uint8Array>): Promise<Buffer> {
-  return new Promise((resolve, reject) => {
-    const chunks: Uint8Array[] = [];
-    const reader = stream.getReader();
-
-    function read() {
-      reader
-        .read()
-        .then(({ done, value }) => {
-          if (done) {
-            resolve(Buffer.concat(chunks.map((c) => Buffer.from(c))));
-            return;
-          }
-          if (value) {
-            chunks.push(value);
-          }
-          read();
-        })
-        .catch(reject);
-    }
-
-    read();
-  });
-}
 
 /**
  * Read the contents of a stored file.
